# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "description": "Stable machine-readable error code from the catalog.",
            "title": "Code",
            "type": "string"
          },
          "invalid_params": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Per-field detail, present only on validation failures.",
            "title": "Invalid Params"
          },
          "key": {
            "description": "Localization key; clients may resolve their own text from it.",
            "title": "Key",
            "type": "string"
          },
          "message": {
            "description": "English message. Localized clients should prefer key + params.",
            "title": "Message",
            "type": "string"
          },
          "params": {
            "additionalProperties": true,
            "description": "Values interpolated into the localized message.",
            "title": "Params",
            "type": "object"
          },
          "severity": {
            "description": "Catalog severity: INFO, WARNING, ERROR or CRITICAL.",
            "title": "Severity",
            "type": "string"
          },
          "trace_id": {
            "description": "Correlates the failure with the audit trail.",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "severity",
          "key",
          "params",
          "message",
          "trace_id"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "ProxyInteractionDetail": {
        "description": "The single-record projection: everything above plus the stored text.\n\nThese four fields are the reason this route is scoped to the interaction's\nowner. They are returned exactly as persisted -- this model describes that\nand does not add, redact or reshape anything.",
        "properties": {
          "behavior_flag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Behavior Flag"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "ISO-8601 UTC with a Z suffix.",
            "title": "Created At"
          },
          "execution_status": {
            "description": "How the proxied call ended, for example `completed`.",
            "title": "Execution Status",
            "type": "string"
          },
          "id": {
            "description": "Row identifier.",
            "title": "Id",
            "type": "string"
          },
          "input_attack_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Attack Type"
          },
          "input_confidence": {
            "title": "Input Confidence",
            "type": "number"
          },
          "input_decision": {
            "description": "Verdict on the prompt: ALLOW, SANITIZE or BLOCK.",
            "title": "Input Decision",
            "type": "string"
          },
          "input_primary_reason": {
            "title": "Input Primary Reason",
            "type": "string"
          },
          "input_raw": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Prompt as received. Null when storage mode did not retain it.",
            "title": "Input Raw"
          },
          "input_sanitized": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Prompt after redaction. Null when nothing was rewritten.",
            "title": "Input Sanitized"
          },
          "input_threats": {
            "items": {
              "type": "string"
            },
            "title": "Input Threats",
            "type": "array"
          },
          "key_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Owning API key, with the internal `key:` prefix stripped. Null for a system record.",
            "title": "Key Id"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "output_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Confidence"
          },
          "output_decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Verdict on the reply. Null when there was no reply to guard.",
            "title": "Output Decision"
          },
          "output_flags": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Output Flags"
          },
          "output_primary_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Primary Reason"
          },
          "output_raw": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Provider reply as received.",
            "title": "Output Raw"
          },
          "output_sanitized": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Reply after redaction.",
            "title": "Output Sanitized"
          },
          "output_threats": {
            "items": {
              "type": "string"
            },
            "title": "Output Threats",
            "type": "array"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Null when the request never reached a provider.",
            "title": "Provider"
          },
          "provider_latency_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Provider time only. Null when no call was made.",
            "title": "Provider Latency Ms"
          },
          "total_latency_ms": {
            "description": "End to end, including detection.",
            "title": "Total Latency Ms",
            "type": "integer"
          },
          "trace_id": {
            "description": "Correlates with the audit trail.",
            "title": "Trace Id",
            "type": "string"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          }
        },
        "required": [
          "id",
          "trace_id",
          "created_at",
          "key_id",
          "user_id",
          "input_decision",
          "input_primary_reason",
          "input_confidence",
          "input_threats",
          "input_attack_type",
          "provider",
          "model",
          "provider_latency_ms",
          "execution_status",
          "output_decision",
          "output_primary_reason",
          "output_confidence",
          "output_threats",
          "behavior_flag",
          "output_flags",
          "total_latency_ms",
          "input_raw",
          "input_sanitized",
          "output_raw",
          "output_sanitized"
        ],
        "title": "ProxyInteractionDetail",
        "type": "object"
      },
      "ProxyInteractionsResponse": {
        "description": "A page of proxy interactions, newest first.\n\nUnlike the audit page, this one echoes `limit` and `offset` back, and both\nare CLAMPED by the handler (limit to 1-200, offset to >= 0) -- so the values\nreturned are the ones actually applied, not the ones requested.",
        "properties": {
          "items": {
            "description": "Newest first; empty when nothing matched.",
            "items": {
              "$ref": "#/components/schemas/ProxyInteraction"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "description": "Page size actually applied after clamping.",
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "description": "Offset actually applied after clamping.",
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "description": "Matching rows across all pages.",
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "limit",
          "offset",
          "items"
        ],
        "title": "ProxyInteractionsResponse",
        "type": "object"
      },
...
      "RecordAttribution": {
        "properties": {
          "app_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "App Id"
          },
          "app_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Resolved application name; null when the lookup found nothing.",
            "title": "App Name"
          },
          "attribution_verified": {
            "title": "Attribution Verified",
            "type": "boolean"
          },
          "dept_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dept Id"
          },
          "dept_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Resolved department name; null when the lookup found nothing.",
            "title": "Dept Name"
          },
          "ip_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ip Address"
          },
          "key_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Key Id"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "tenant_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Id"
          },
          "user_agent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Agent"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          }
        },
        "required": [
          "tenant_id",
          "dept_id",
          "dept_name",
          "app_id",
          "app_name",
          "source",
          "user_id",
          "key_id",
          "ip_address",
          "user_agent",
          "attribution_verified"
        ],
        "title": "RecordAttribution",
        "type": "object"
      },
      "RecordProcessing": {
        "properties": {
          "detection_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detection Mode"
          },
          "execution_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Mode"
          },
          "latency_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms"
          },
          "llm_invoked": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Invoked"
          },
          "policy_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Which policy layer resolved the decision, or `cache` for a cache hit.",
            "title": "Policy Source"
          }
        },
        "required": [
          "latency_ms",
          "llm_invoked",
          "detection_mode",
          "execution_mode",
          "policy_source"
        ],
        "title": "RecordProcessing",
        "type": "object"
      },
...
      "RequestRecordResponse": {
        "properties": {
          "attribution": {
            "$ref": "#/components/schemas/RecordAttribution"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "confidence_band": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Band"
          },
          "decision": {
            "title": "Decision",
            "type": "string"
          },
          "detection_scores": {
            "additionalProperties": {
              "type": "number"
            },
            "description": "Per-detector scores as persisted. EMPTY for a caller that may not read layer scores -- the key stays present so consumers need no special case.",
            "title": "Detection Scores",
            "type": "object"
          },
          "execution_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Mode"
          },
          "guardrail_scores": {
            "additionalProperties": {
              "type": "number"
            },
            "description": "Guardrail scores as persisted. Empty under the same restriction as detection_scores.",
            "title": "Guardrail Scores",
            "type": "object"
          },
          "input_hash": {
            "description": "Hash of the scanned input; the input itself is never stored here.",
            "title": "Input Hash",
            "type": "string"
          },
          "input_length": {
            "title": "Input Length",
            "type": "integer"
          },
          "input_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Declared provenance, for example `user_prompt` or `retrieved_document`.",
            "title": "Input Source"
          },
          "is_proxy": {
            "title": "Is Proxy",
            "type": "boolean"
          },
          "primary_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Reason"
          },
          "processing": {
            "$ref": "#/components/schemas/RecordProcessing"
          },
          "proxy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RecordProxyDetail"
              },
              {
                "type": "null"
              }
            ],
            "description": "Proxy lifecycle detail. Absent for a scan-only request."
          },
          "risk_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Score"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Caller-supplied agent-run correlation; never an authorization input.",
            "title": "Run Id"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "severity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "INFO, WARNING, ERROR or CRITICAL.",
            "title": "Severity"
          },
          "threats": {
            "items": {
              "type": "string"
            },
            "title": "Threats",
            "type": "array"
          },
          "timestamp": {
            "description": "ISO-8601 UTC with a Z suffix.",
            "title": "Timestamp",
            "type": "string"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          },
          "turn_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Turn Index"
          }
        },
        "required": [
          "trace_id",
          "timestamp",
          "execution_mode",
          "is_proxy",
          "severity",
          "attribution",
          "decision",
          "risk_score",
          "primary_reason",
          "confidence",
          "confidence_band",
          "threats",
          "input_hash",
          "input_length",
          "run_id",
          "session_id",
          "turn_index",
          "input_source",
          "detection_scores",
          "guardrail_scores",
          "processing"
        ],
        "title": "RequestRecordResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "AI Security Gateway for Intelligent Applications",
    "title": "WrapSec",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/ai/requests/{trace_id}": {
      "get": {
        "description": "Returns the full audit record for a single request by trace_id.\nAdmin keys use an unscoped lookup; all other keys are scoped to their dept_id.\nFor proxy requests, the response is enriched with the full proxy_interactions record\nincluding provider response, output decision, and execution status.\n404 if the record does not exist or is out of scope.",
        "operationId": "get_request_v1_ai_requests__trace_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "trace_id",
            "required": true,
            "schema": {
              "title": "Trace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestRecordResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Missing or invalid credentials."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No such trace_id, or it belongs to another scope."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed. Published for every parameterized route; this one validates nothing, so it is not reachable here."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Rate limit exceeded: this route shares the global `/v1/ai` bucket."
          }
        },
        "summary": "Get Request",
        "tags": [
          "Gateway"
        ]
      }
    },
...
    "/v1/proxy/interactions": {
      "get": {
        "operationId": "list_proxy_interactions_v1_proxy_interactions_get",
        "parameters": [
          {
            "in": "query",
            "name": "execution_status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Execution Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProxyInteractionsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Missing or invalid credentials."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A query parameter could not be parsed."
          }
        },
        "summary": "List Proxy Interactions",
        "tags": [
          "Proxy"
        ]
      }
    },
    "/v1/proxy/interactions/{trace_id}": {
      "get": {
        "operationId": "get_proxy_interaction_v1_proxy_interactions__trace_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "trace_id",
            "required": true,
            "schema": {
              "title": "Trace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProxyInteractionDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Missing or invalid credentials."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No interaction with this trace_id, or it is out of the caller's scope. The two are deliberately indistinguishable."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A query parameter could not be parsed."
          }
        },
        "summary": "Get Proxy Interaction",
        "tags": [
          "Proxy"
        ]
      }
    },
...
  }
}